requirements.txt
langchain
langchain_experimental
langchain-google-genai
streamlit
pandas
python-environ
tabulate